requirements.txt
Flask==2.3.3
flask-cors==4.0.0
Werkzeug==2.3.7
PyJWT==2.8.0
Flask-Mail==0.9.1
itsdangerous==2.1.2